+++ revolv/payments/utils.py
@@ -1,0 +1,2 @@
+class NotEnoughFundingException(Exception):
+    pass
